feat: Renamed files to fit captive portal setup. Added mock website
Added the mock website for the current year. Heavily edited the HTML to make the links that we want work and simplify the use of the website. Also had to rename Celia's files to make it work properly. This was because the portal plugin setups up a framework that we have to adhere to.